requirements.txt
discord.py
reactionmenu
sqlalchemy
humanize
pynacl
yt_dlp
validators
requests
types-requests
rebelai
